Posix: Replace SDL 1.2 with SDL2
This is required for DeSmuME to start without X11, for instance on a
pure Wayland system.

+++ desmume/src/frontend/posix/shared/ctrlssdl.h
@@ -100,7 +100,7 @@ struct ctrls_event_config {
 #endif
 };
 
-void load_default_config(const u16 kbCfg[]);
+void load_default_config(const u32 kbCfg[]);
 BOOL init_joy( void);
 void uninit_joy( void);
 u16 get_joy_key(int index);
